requirements.txt
openai>=1.50.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.1

# Admin portal
streamlit>=1.40.0
pandas>=2.2.0
xlsxwriter>=3.2.0

# Document parsing
pdfplumber>=0.11.0
python-docx>=1.1.0
openpyxl>=3.1.2

# Chat API
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
